Automata

Types of Automata

newtype DTransition q a

Constructors

DTransition (Map (q, a) q)

newtype NTransitionE q a

Constructors

NTransitionE (Map (q, Maybe a) (Set q))
with epsilon-transitions

newtype NTransition q a

Constructors

NTransition (Map (q, a) (Set q))

newtype ATransition q a

Constructors

ATransition (Map (q, a) (BasicPropLogic q))
should be positive boolean formula. Todo: write function to enforce this.

data FA q a t

the general type of a finite automaton.

Constructors

FA

Fields

alphabet :: Set a

states :: Set q

init :: q

just one starting state.

transition :: t q a

final :: Set q

type DFA q a = FA q a DTransition

type NFAE q a

Arguments

= FA q a NTransitionE
NFAE with alphabet as has transition function from (as + epsilon) x states -> ...

type NFA q a = FA q a NTransition

type AFA q a = FA q a ATransition

isFinal :: Ord q => q -> Set q -> Bool

Easy Translations (into more succint automata)

d2nE :: (Eq a, Ord q) => DFA q a -> NFAE q a

d2n :: Ord q => DFA q a -> NFA q a

n2a :: (Eq a, Ord q) => NFA q a -> AFA q a

bigOr :: (Ord p, Ord u, Ord b) => Set p -> PropLogic p u b

Harder Translations (compiling into lower level automata)

NFAs to DFAs

nE2n :: NFAE q a -> NFA q a

n2d :: (Ord a, Ord q) => NFA q a -> DFA (Set q) a

post :: (Ord a, Ord q) => Set q -> Map (q, a) (Set r) -> a -> Set r

states reachable from set s according to t, via action a

subsets :: Ord a => Set a -> Set (Set a)

powerset

AFAs to NFAs

a2n :: (Ord a, Ord q) => AFA q a -> NFA (Set q) a

minSat :: forall q. Ord q => Set q -> (Set q -> Bool) -> Set (Set q)

LDL to AFAs

ldl2afa :: LDLogic p -> AFA (LDLogic p) (Set p)

Emptiness testing

dfaEmpty :: DFA q a -> Bool