Logic

Documentation

data PropLogic p u b

PropLogic is a data type for logics based on propositional logic- particularly modal logics.

Constructors

PropConst Bool

PropVar p

Not (PropLogic p u b)

And (PropLogic p u b) (PropLogic p u b)

Or (PropLogic p u b) (PropLogic p u b)

Unary u (PropLogic p u b)

Binary b (PropLogic p u b) (PropLogic p u b)

foldLog :: (Bool -> r) -> (p -> r) -> (r -> r) -> (r -> r -> r) -> (r -> r -> r) -> (u -> r -> r) -> (b -> r -> r -> r) -> PropLogic p u b -> r

mapLog :: (Bool -> Bool) -> (p -> q) -> PropLogic p u b -> PropLogic q u b

satisfyBasicProp :: (p -> Bool) -> BasicPropLogic p -> Bool

connect :: String -> String -> String -> String

logic_pretty_print :: Show p => (u -> String -> String) -> (b -> String -> String -> String) -> PropLogic p u b -> String

logic_pretty_print_ :: (u -> String -> String) -> (b -> String -> String -> String) -> PropLogic String u b -> String

data Void

type BasicPropLogic p = PropLogic p Void Void

data UnaryLTL

Constructors

Next

data BinaryLTL

Constructors

Until

type LTLogic p = PropLogic p UnaryLTL BinaryLTL

type LDLogic p = PropLogic p (Prog p) Void

data Prog p

Constructors

Base (BasicPropLogic p)

Test (LDLogic p)

Plus (Prog p) (Prog p)

Comp (Prog p) (Prog p)

Star (Prog p)

foldProg :: (BasicPropLogic p -> r) -> (LDLogic p -> r) -> (r -> r -> r) -> (r -> r -> r) -> (r -> r) -> Prog p -> r

trueProg :: Prog p

diamond :: Prog p -> LDLogic p -> LDLogic p

diamond p is a unary modality that transforms LDL formulae. e.g. true f where f is an LDL formula, is diamond true f. square is its dual.

square :: Prog p -> LDLogic p -> LDLogic p

ltl2ldl :: LTLogic p -> LDLogic p